+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed bug that allocate a new qubit when passing an empty list to a quant
 - Fixed error raised by repeated basis states in `dump.show`.
+- Fixed import error raised by `lib.w`.
 
 ## 0.3.1